requirements.txt
# Streamlit for the UI
streamlit==1.34.0
# Assembly AI for transcription
requests==2.31.0
# PDF generation
fpdf==1.7.2
# Gemini API
google-generativeai==0.3.2
